requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pandas==2.2.3
chromadb==0.5.11
google-generativeai==0.8.3
